refactor: :building_construction: add AddButton component
