+++ src/App.tsx
@@ -3,27 +3,8 @@ import Group from "./Group";
 import Task from "./Task";
 
 import "./App.css";
-import Confetti, { ConfettiRef } from "./Confetti";
-
-const plugSVG = (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    className="h-6 w-6"
-    fill="none"
-    viewBox="0 0 24 24"
-    stroke="currentColor"
-  >
-    <path
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={2}
-      d="M12 6v6m0 0v6m0-6h6m-6 0H6"
-    />
-  </svg>
-);
-
-const buttonStyle =
-  "flex items-center justify-center bg-slate-500 rounded p-2 opacity-30 hover:opacity-70";
+import Confetti, { ConfettiRef } from "./components/Confetti";
+import AddButton from "./components/AddButton";
 
 function App() {
   const [groups, setGroup] = useState<Group[]>([
@@ -142,16 +123,7 @@ function App() {
                 </button>
               );
           })}
-          {add && (
-            <button
-              className={buttonStyle}
-              onClick={() => {
-                addNewTask(group);
-              }}
-            >
-              {plugSVG}
-            </button>
-          )}
+          {add && <AddButton onClick={() => addNewTask(group)} />}
         </div>
         {group.length === 0 && add === false && (
           <h2 className="text-3xl">Let's do something.</h2>
@@ -182,14 +154,10 @@ function App() {
             add={true}
           />
         ))}
-        <button
-          className={"absolute left-full ml-4 h-full w-10" + buttonStyle}
-          onClick={() => {
-            setGroup([...groups, new Group("New Group")]);
-          }}
-        >
-          {plugSVG}
-        </button>
+        <AddButton
+          className="absolute left-full ml-4 h-full w-10"
+          onClick={() => setGroup([...groups, new Group("New Group")])}
+        />
       </main>
     </div>
   );
